+++ daemon/src/actors/window.rs
@@ -1,13 +1,12 @@
 use bytes::Bytes;
 use tokio::sync::mpsc;
-use tracing::trace;
 
 use crate::{
     actors::{
         pane::{Pane, PaneHandle},
         session::SessionHandle,
     },
-    error::Result,
+    prelude::*,
 };
 
 #[derive(Debug)]
@@ -17,6 +16,7 @@ pub enum WindowEvent {
     Redraw,
     Kill,
 }
+#[allow(unused)]
 #[derive(Debug)]
 pub enum WindowState {
     Focused,
@@ -28,62 +28,61 @@ pub struct Window {
     handle: WindowHandle,
     rx: mpsc::Receiver<WindowEvent>,
     pane_handle: PaneHandle, // TODO: handle more panes
+    #[allow(unused)]
     window_state: WindowState,
 }
 impl Window {
-    async fn handle_user_input(&mut self, bytes: Bytes) {
-        self.pane_handle.send_user_input(bytes).await.unwrap()
+    async fn handle_user_input(&mut self, bytes: Bytes) -> Result<()> {
+        self.pane_handle.send_user_input(bytes).await
     }
-    async fn handle_pane_output(&mut self, bytes: Bytes) {
-        self.session_handle.send_window_output(bytes).await;
+    async fn handle_pane_output(&mut self, bytes: Bytes) -> Result<()> {
+        self.session_handle.send_window_output(bytes).await
     }
-    async fn handle_redraw(&mut self) {
-        self.pane_handle.rerender().await.unwrap()
+    async fn handle_redraw(&mut self) -> Result<()> {
+        self.pane_handle.request_rerender().await
     }
 }
 impl Window {
     pub fn spawn(session_handle: SessionHandle) -> Result<WindowHandle> {
-        let window = Window::new(session_handle);
+        let window = Window::new(session_handle)?;
         window.run()
     }
-    fn new(session_handle: SessionHandle) -> Self {
+    fn new(session_handle: SessionHandle) -> Result<Self> {
         let (tx, rx) = mpsc::channel(10);
         let handle = WindowHandle { tx };
-        let pane_handle = Pane::spawn(handle.clone()).unwrap();
-        Self {
+        let pane_handle = Pane::spawn(handle.clone())?;
+        Ok(Self {
             session_handle,
             handle,
             rx,
             pane_handle,
             window_state: WindowState::Focused,
-        }
+        })
     }
     fn run(mut self) -> crate::error::Result<WindowHandle> {
         let handle_clone = self.handle.clone();
         let _task = tokio::spawn({
             async move {
                 loop {
-                    tokio::select! {
-                        Some(event) = self.rx.recv() => {
-                            use WindowEvent::*;
-                            match event {
-                                UserInput(bytes) => {
-                                    trace!("Window: UserInput");
-                                    self.handle_user_input(bytes).await;
-                                },
-                                PaneOutput(bytes) => {
-                                    trace!("Window: PaneOutput");
-                                    self.handle_pane_output(bytes).await;
-                                }
-                                Redraw => {
-                                    trace!("Window: Redraw");
-                                    self.pane_handle.rerender().await.unwrap();
-                                }
-                                Kill => {
-                                    trace!("Window: Kill");
-                                    self.pane_handle.kill().await.unwrap();
-                                    break;
-                                }
+                    if let Some(event) = self.rx.recv().await {
+                        use WindowEvent::*;
+                        match event {
+                            UserInput(bytes) => {
+                                trace!("Window: UserInput");
+                                self.handle_user_input(bytes).await.unwrap();
+                            }
+                            PaneOutput(bytes) => {
+                                trace!("Window: PaneOutput");
+                                self.handle_pane_output(bytes).await.unwrap();
+                            }
+                            Redraw => {
+                                trace!("Window: Redraw");
+                                self.handle_redraw().await.unwrap();
+                            }
+                            Kill => {
+                                trace!("Window: Kill");
+                                self.pane_handle.kill().await.unwrap();
+                                break;
                             }
                         }
                     }
@@ -99,22 +98,16 @@ pub struct WindowHandle {
     tx: mpsc::Sender<WindowEvent>,
 }
 impl WindowHandle {
-    pub async fn send_pane_output(&self, bytes: Bytes) {
-        self.tx.send(WindowEvent::PaneOutput(bytes)).await.unwrap();
+    pub async fn send_pane_output(&self, bytes: Bytes) -> Result<()> {
+        Ok(self.tx.send(WindowEvent::PaneOutput(bytes)).await?)
     }
-    pub async fn send_user_input(&self, bytes: Bytes) {
-        self.tx.send(WindowEvent::UserInput(bytes)).await.unwrap();
+    pub async fn send_user_input(&self, bytes: Bytes) -> Result<()> {
+        Ok(self.tx.send(WindowEvent::UserInput(bytes)).await?)
     }
-
-    pub async fn redraw(&self) {
-        self.tx.send(WindowEvent::Redraw).await.unwrap();
+    pub async fn redraw(&self) -> Result<()> {
+        Ok(self.tx.send(WindowEvent::Redraw).await?)
     }
-    pub async fn kill(&self) -> crate::error::Result<()> {
-        self.tx.send(WindowEvent::Kill).await.unwrap();
-        Ok(())
-    }
-
-    fn is_alive(&self) -> bool {
-        !self.tx.is_closed()
+    pub async fn kill(&self) -> Result<()> {
+        Ok(self.tx.send(WindowEvent::Kill).await?)
     }
 }
